imessage applied

+++ TwitterAPIDemo/TwitterAPIDemo.Android/Message_droid.cs
@@ -24,7 +24,7 @@ namespace TwitterAPIDemo.Droid
 
         public void Shorttime(string message)
         {
-            Toast.MakeText(Android.App.Application.Context, message, ToastLength.Long).Show();
+            Toast.MakeText(Android.App.Application.Context, message, ToastLength.Short).Show();
         }
     }
 }
